+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## [1.0.4] - 06-Jun-2021.
+
+- String methods added
+
+## [1.0.3] - 06-Jun-2021.
+
+- Static methods
+
 ## [1.0.2] - 28-Mar-2021.
 
 - New methods added